Jangan otak-atik master dulu deh. Edit branch yang temp aja.

+++ PA_13_Majestic_FlapFlap/PA_13_Majestic_FlapFlap/TPoint.vb
@@ -1,5 +1,5 @@
 ﻿Public Class TPoint
-    Public X, Y, Z, W As Integer 'Store the x,y and z coord of this vertex as well as its index number + w for perspective
+    Public X, Y, Z, W As Single 'Store the x,y and z coord of this vertex as well as its index number + w for perspective
     'Public PointIndex As Integer = -1
 
     Public Sub New()
@@ -10,7 +10,7 @@
         '   Me.PointIndex = 0
     End Sub
 
-    Public Sub New(X As Integer, Y As Integer, Z As Integer)
+    Public Sub New(X As Single, Y As Single, Z As Single)
         Me.X = X
         Me.Y = Y
         Me.Z = Z
@@ -18,7 +18,7 @@
         '  Me.PointIndex = 0
     End Sub
 
-    Public Sub SetCoordinates(ByVal X As Integer, ByVal Y As Integer, ByVal Z As Integer) 'Method to set coordinates
+    Public Sub SetCoordinates(ByVal X As Single, ByVal Y As Single, ByVal Z As Single) 'Method to set coordinates
         Me.X = X
         Me.Y = Y
         Me.Z = Z
